Fix for an unrealpak crash when a compressed files size doesn't fit in a int32
#rb none
#jira UE-97132

#ushell-cherrypick of 14260544 by Graeme.Thornton
#lockdown marcus.wassmer

[CL 15282116 by Martin Sevigny in 4.26 branch]

+++ Engine/Source/Developer/PakFileUtilities/Private/PakFileUtilities.cpp
@@ -217,7 +217,7 @@ class FMemoryCompressor
 {
 public:
 	/** Divide into blocks and start compress asynchronously */
-	FMemoryCompressor(uint8* UncompressedBuffer, int32 UncompressedSize, FName Format, int32 CompressionBlockSize) :
+	FMemoryCompressor(uint8* UncompressedBuffer, int64 UncompressedSize, FName Format, int32 CompressionBlockSize) :
 		Index(0)
 			{
 		// Divide into blocks and start compression async tasks.
